+++ src/ell/__init__.py
@@ -1,6 +1,6 @@
-from ell.lmp.text import simple
+from ell.lmp.simple import simple
 from ell.lmp.tool import tool
-from ell.lmp.multimodal import complex
+from ell.lmp.complex import complex
 from ell.types.message import Message, ContentBlock, system, user, assistant
 from ell.__version__ import __version__
